refactor(parse): embed sequential parse step into a main application and execute it from there
--date=0.day.ago
