+++ score.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import pandas as pd
 from tqdm import tqdm as tqdm
 
-import global_options
+import app_corporate_culture_measurer.global_options as global_options
 from culture import culture_dictionary, file_util
 
 # @TODO: The scoring functions are not memory friendly. The entire pocessed corpus needs to fit in the RAM. Rewrite a memory friendly version.
