+++ src/app_corporate_culture_measurer/app_corporate_culture_measurer.py
@@ -1,3 +1,10 @@
 import logging
 import sys
 from logging import log
+
+
+from preprocess.parse import sequential_parse
+
+
+if __name__ == "__main__":
+    sequential_parse()
