+++ src/app_corporate_culture_measurer/culture/culture_dictionary.py
