+++ src/app_corporate_culture_measurer/culture/file_util.py
