+++ src/app_corporate_culture_measurer/data/input/id2firms.csv
