+++ src/app_corporate_culture_measurer/global_options.py
