+++ src/app_corporate_culture_measurer/preprocess/parse.py
@@ -2,24 +2,26 @@ import datetime
 import itertools
 import os
 from pathlib import Path
-
+from functools import partial
 from stanfordnlp.server import CoreNLPClient
 
-import global_options
+import global_options as global_options
 from culture import file_util, preprocess
 
 
-def process_line(line, lineID):
-    """Process each line and return a tuple of sentences, sentence_IDs, 
-    
+def process_line(corpus_preprocessor, line, lineID):
+    """Process each line and return a tuple of sentences, sentence_IDs,
+
     Arguments:
-        line {str} -- a document 
+        line {str} -- a document
         lineID {str} -- the document ID
-    
+
     Returns:
-        str, str -- processed document with each sentence in a line, 
+        str, str -- processed document with each sentence in a line,
                     sentence IDs with each in its own line: lineID_0 lineID_1 ...
     """
+    sentences_processed = []
+    doc_sent_ids = []
     try:
         sentences_processed, doc_sent_ids = corpus_preprocessor.process_document(
             line, lineID
@@ -39,7 +41,7 @@ def process_largefile(
     chunk_size=100,
     start_index=None,
 ):
-    """ A helper function that transforms an input file + a list of IDs of each line (documents + document_IDs) to two output files (processed documents + processed document IDs) by calling function_name on chunks of the input files. Each document can be decomposed into multiple processed documents (e.g. sentences). 
+    """A helper function that transforms an input file + a list of IDs of each line (documents + document_IDs) to two output files (processed documents + processed document IDs) by calling function_name on chunks of the input files. Each document can be decomposed into multiple processed documents (e.g. sentences).
     Supports parallel with Pool.
 
     Arguments:
@@ -92,6 +94,7 @@ def process_largefile(
             ):
                 output_lines.append(output_line)
                 output_line_ids.append(output_line_id)
+
             output_lines = "\n".join(output_lines) + "\n"
             output_line_ids = "\n".join(output_line_ids) + "\n"
             with open(output_file, "a", newline="\n") as f_out:
@@ -101,7 +104,7 @@ def process_largefile(
                     f_out.write(output_line_ids)
 
 
-if __name__ == "__main__":
+def sequential_parse():
     with CoreNLPClient(
         properties={
             "ner.applyFineGrained": "false",
@@ -113,9 +116,9 @@ if __name__ == "__main__":
         max_char_length=1000000,
     ) as client:
         corpus_preprocessor = preprocess.preprocessor(client)
-        in_file = Path(global_options.DATA_FOLDER, "input", "documents.txt")
+        in_file = Path(global_options.DATA_FOLDER, "input_for_test", "documents.txt")
         in_file_index = file_util.file_to_list(
-            Path(global_options.DATA_FOLDER, "input", "document_ids.txt")
+            Path(global_options.DATA_FOLDER, "input_for_test", "document_ids.txt")
         )
         out_file = Path(
             global_options.DATA_FOLDER, "processed", "parsed", "documents.txt"
@@ -128,6 +131,8 @@ if __name__ == "__main__":
             output_file=out_file,
             input_file_ids=in_file_index,
             output_index_file=output_index_file,
-            function_name=process_line,
+            function_name=partial(
+                process_line, corpus_preprocessor
+            ),  # NB: I am using a Partial function for now! I will remove later with an OO design
             chunk_size=global_options.PARSE_CHUNK_SIZE,
         )
