+++ src/app_corporate_culture_measurer/resources/StopWords_Generic.txt
