+++ start_system.sh
@@ -1,19 +1,21 @@
+echo "Starting up the sytem."
+cd ./src/app_corporate_culture_measurer/
 echo "parse"
-    python parse.py
+    python app_corporate_culture_measurer.py
 echo "done"
 
-echo "clean and train"
-    python clean_and_train.py
-echo "done"
+# echo "clean and train"
+#     python clean_and_train.py
+# echo "done"
 
-echo "create culture dictionary"
-    python create_dict.py
-echo "done"
+# echo "create culture dictionary"
+#     python create_dict.py
+# echo "done"
 
-echo "score"
-    python score.py
-echo "done"
+# echo "score"
+#     python score.py
+# echo "done"
 
-echo "aggregate firms"
-    python aggregate_firms.py
-echo "done"
+# echo "aggregate firms"
+#     python aggregate_firms.py
+# echo "done"
